feat: track decorator reform
RFC for this change: salesforce/lwc-rfcs#4

The logic for this PR is as follows:

When the compiler compiles the class, it can extract any field that is not decorated with `@api`, `@track` or `@wire`, and pass the metadata through the registerComponent call.

For every class field that is passed to the registerComponent, we will create a getter and a setter in the prototype in order to observe mutations in to class field.

No change is needed to the logic in the engine.